Move legacy transit only code from euclidean RWH to new class
